Brand create

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\DataTables\CategoryDataTable;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Str;
 
 
@@ -46,8 +47,8 @@ class CategoryController extends Controller
         $category->status = $request->status;
         $category->save();
 
-        flash()->success('Created successfully.');
-        return redirect()->route('admin.category.create');
+        toastr()->success('Created successfully.');
+        return redirect()->route('admin.category.index');
 
     }
 
@@ -97,6 +98,10 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
+        $subCategory = SubCategory::where('category_id', $category->id)->count();
+        if($subCategory > 0){
+            return response(['status' => 'error', 'message' => 'This item contain sub-items, for delete this you have to delete the sub-items first!']);
+        }
         $category->delete();
         return response(['status' => 'success', 'message' => 'Delete Successfully!']);
     }
